Simplified disableRotation
The getRotation() and getOrientation() always have the same result, so the elses are never executed. This led to fixed portrait orientation if the orientation was 0° and fixed 90° landscape otherwise.

However, screenOrientation has a constant that can be used to fix the display in its current orientation, regardless of what it is.